[phsa] Disabled sync_2way.cpp as it is really difficult to make work reliable on stock HSA.
This is because of the forever loop inside a work-item. HSA does
not guarantee forward progress in this case unless there's a barrier.

The one injected there causes a race condition: some of the work-items
might read value 16 thus exit the loop and not reach the barrier in
the next iteration, thus the barrier condition of 'all or none' won't
fulfill, leading to undefined behavior. Fixing this would require
monitoring work-group wide state of the maximum seen value or similar.